pyproject.toml
[project]
name = "kwatcher"
version = "0.2.0"
description = "Speech-to-text Telegram Python bot"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=22.3,<23.0)",
    "vosk (>=0.3.45,<0.4.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

